requirements.txt
PyYAML>=6.0
numpy>=1.24
scipy>=1.10

